fix(get_trame_versions): avoid errors
For some currently unknown reason, not every PathDistribution contains a "Name" in the metadata.
These are probably not trame packages and thus we can probably just ignore them. But if we notice
that we are missing trame packages in the future, we should potentially find an alternative way to
identify them.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>